% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_add_transboundary_pa.R
\name{ec_add_transboundary_pa}
\alias{ec_add_transboundary_pa}
\title{Add Transboundary Protected Areas to Analysis}
\usage{
ec_add_transboundary_pa(prepared_data, buffer_km = 100)
}
\arguments{
\item{prepared_data}{List object returned by \code{\link[=ec_prepare_data]{ec_prepare_data()}} containing processed
protected areas and ecoregion data}

\item{buffer_km}{Numeric. Buffer distance in kilometers around the protected area
to search for transboundary protected areas. Default is 100 km}
}
\value{
An sf object containing both ecoregion and transboundary protected areas with
a 'type' column indicating whether each PA is within the ecoregion ("ecoregion") or
outside it ("transboundary")
}
\description{
This function identifies and includes protected areas located outside an ecoregion boundary
that may contribute to connectivity between protected areas within the ecoregion by acting
as stepping stones or corridors.
}
\examples{
\donttest{
library(dplyr)
library(sf)

# Benin - treated as single ecoregion
benin <- fiwac \%>\%
  dplyr::filter(ISO3_CODE == "BEN")

# repare data first
prepared_data <- ec_prepare_data(
  pa = pas,
  ecoregion = benin,
  min_area = 1
)

# Add transboundary PAs with default 70km buffer
complete_pa <- ec_add_transboundary_pa(prepared_data, buffer_km = 70)
head(complete_pa, 5)
}

}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017).
Protected areas in the world's ecoregions: How well connected are they?
Ecological Indicators, 76, 144-158. \doi{10.1016/j.ecolind.2016.12.047}
}
