% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_protconn.R
\name{ec_protconn}
\alias{ec_protconn}
\title{Calculate the Protected Connected Indicator}
\usage{
ec_protconn(
  pa,
  max_prob_matrix,
  include_fraction = FALSE,
  prepared_data = NULL,
  ecoregion_area
)
}
\arguments{
\item{pa}{An \code{sf} object of protected areas. Must include a \code{type} column
identifying \code{"transboundary"} PAs as retuned by \code{\link[=ec_add_transboundary_pa]{ec_add_transboundary_pa()}}.}

\item{max_prob_matrix}{Numeric matrix of maximum product probabilities between all
protected areas, typically generated by \code{\link[=ec_max_product_probability]{ec_max_product_probability()}}}

\item{include_fraction}{Logical. If \code{TRUE}, ProtConn fractions are also
calculated following Saura et al. (2017). Default: \code{FALSE}.}

\item{prepared_data}{Optional. A prepared dataset (see \code{\link[=ec_prepare_data]{ec_prepare_data()}})
containing ecoregions and PA information. Required if \code{include_fraction = TRUE}.}

\item{ecoregion_area}{sf object or total area of the ecoregion in(\eqn{\text{km}^2})}
}
\value{
A tibble with the following indicators (all expressed as percentages):
\describe{
\item{ProtConn}{Protected Connected indicator: percentage of the ecoregion
that is protected and connected.}
\item{Prot}{Protected area coverage: percentage of the ecoregion that is
protected, regardless of connectivity.}
\item{ProtUnconn}{Protected unconnected: the difference between coverage
(Prot) and connected coverage (ProtConn).}
\item{RelConn}{Relative connectivity: proportion of protected area that is
connected, i.e. \verb{ProtConn / Prot × 100}.}
}

If \code{include_fraction = TRUE}, additional columns are included that decompose
ProtConn into fractions (intra-PA connectivity, connectivity through direct
links, connectivity through transboundary PAs, etc.).
}
\description{
The Protected Connected (ProtConn) indicator quantifies the percentage of
a given ecoregion that is both protected and functionally connected,
considering species' dispersal abilities and potential transboundary linkages
between protected areas (PAs). It extends the concept of protected area (PA)
coverage by incorporating connectivity into conservation effectiveness
assessments.
}
\details{
ProtConn is based on the Equivalent Connected Area (ECA) concept (Saura et al.
2011; Saura et al. 2017). The ECA integrates PA area and dispersal
probabilities into a measure of the amount of habitat that is both protected
and connected. ProtConn expresses ECA as a percentage of the ecoregion area,
enabling comparisons across regions of different size.

Formally:
\deqn{ProtConn = 100 \times \frac{ECA}{A_{L}}}
where:
\itemize{
\item \eqn{ECA} is the Equivalent Connected Area (\eqn{\text{km}^2}),
\item \eqn{A_{L}} is the total area of the ecoregion (\eqn{\text{km}^2}).
}
}
\examples{
\donttest{
library(dplyr)
library(sf)
# Benin - treated as single ecoregion
benin <- fiwac \%>\%
  dplyr::filter(ISO3_CODE == "BEN")

# Protected area
prot_area <- pas

# repare data first
prepared_data <- ec_prepare_data(
  pa = prot_area,
  ecoregion = benin,
  min_area = 1
)
# Add transboundary PAs with default 70km buffer
complete_pa <- ec_add_transboundary_pa(prepared_data, buffer_km = 70)

# Calculate edge-to-edge distances
pair_dist <- ec_pa_distances(complete_pa)
prob_matrix <- ec_dispersal_probability(pair_dist, median_dispersal = 10)
pstar <- ec_max_product_probability(prob_matrix)

ec_protconn(pa = complete_pa,
            max_prob_matrix = pstar,
            ecoregion_area = benin)
}

}
\references{
Saura, S., Bastin, L., Battistella, L., Mandrici, A., & Dubois, G. (2017).
Protected areas in the world's ecoregions: How well connected are they?
Ecological Indicators, 76, 144-158. \doi{10.1016/j.ecolind.2016.12.047}

Saura, S., & Pascual-Hortal, L. (2007). A new habitat availability index to integrate
connectivity in landscape conservation planning: comparison with existing indices and
application to a case study. Landscape and Urban Planning, 83(2-3), 91-103.
\doi{doi.org/10.1016/j.landurbplan.2007.03.005}
}
