% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_protconn_fractions.R
\name{ec_protconn_fractions}
\alias{ec_protconn_fractions}
\title{Calculate ProtConn Fraction Components}
\usage{
ec_protconn_fractions(
  pa,
  prepared_data,
  max_prob_matrix,
  ecoregion_area,
  protconn
)
}
\arguments{
\item{pa}{An \code{sf} object of protected areas as returned by \code{\link[=ec_add_transboundary_pa]{ec_add_transboundary_pa()}}.}

\item{prepared_data}{A list as returned by preprocessing functions \code{\link[=ec_prepare_data]{ec_prepare_data()}}.}

\item{max_prob_matrix}{Numeric matrix of maximum product probabilities between all
protected areas, typically generated by \code{\link[=ec_max_product_probability]{ec_max_product_probability()}}}

\item{ecoregion_area}{sf object or total area of the ecoregion in(\eqn{\text{km}^2})}

\item{protconn}{Numeric. Overall ProtConn value as calculated by \code{\link[=ec_protconn]{ec_protconn()}}.}
}
\value{
A tibble with the following ProtConn fraction components (percentages of ProtConn):
\itemize{
\item \code{ProtConn_Prot} – contribution of movement restricted to protected lands.
\item \code{ProtConn_Within} – contribution of movements confined within individual PAs.
\item \code{ProtConn_Contig} – contribution of movements across contiguous (adjacent) PAs.
\item \code{ProtConn_Unprot} – contribution of movements requiring unprotected lands.
\item \code{ProtConn_Trans} – contribution of transboundary linkages between ecoregions
\item \code{Correction_Factor} – adjustment factor applied to avoid overestimation
due to transboundary PA aggregation - double counting.
}
}
\description{
Decomposes the overall Protected Connected Indicator (ProtConn; Saura et al. 2017)
into fractions that quantify the relative contribution of different connectivity
pathways: movement within individual protected areas, through contiguous protected
areas, through unprotected lands, and across international boundaries.
}
\examples{
\donttest{
library(dplyr)
library(sf)
# Benin - treated as single ecoregion
benin <- fiwac \%>\%
  dplyr::filter(ISO3_CODE == "BEN")

# Protected area
prot_area <- pas

# repare data first
prepared_data <- ec_prepare_data(
  pa = prot_area,
  ecoregion = benin,
  min_area = 1
)
# Add transboundary PAs with default 70km buffer
complete_pa <- ec_add_transboundary_pa(prepared_data, buffer_km = 70)

# Calculate edge-to-edge distances
pair_dist <- ec_pa_distances(complete_pa)
prob_matrix <- ec_dispersal_probability(pair_dist, median_dispersal = 10)
pstar <- ec_max_product_probability(prob_matrix)

# Calculate protconn fractions
ec_protconn_fractions(pa = complete_pa,
                      prepared_data = prepared_data,
                      max_prob_matrix = prob_matrix,
                      ecoregion_area = benin,
                      protconn = 14)
}

}
\references{
Saura, S., Bastin, L., Battistella, L. et al. (2017).
Protected areas in the world’s ecoregions: How well connected are they?
Ecological Indicators, 76, 144–158. \doi{10.1016/j.ecolind.2016.12.047}

Saura, S. & Rubio, L. (2010).
A common currency for the different ways in which patches and links can
contribute to habitat availability and connectivity in the landscape.
Ecography, 33, 523–537. \doi{10.1111/j.1600-0587.2009.05760.x}
}
\seealso{
\code{\link[=ec_protconn]{ec_protconn()}}, \code{\link[=ec_eca]{ec_eca()}}, \code{\link[=ec_max_product_probability]{ec_max_product_probability()}}
}
